+++ R/igraph.r
@@ -20,36 +20,37 @@ NULL
 #' @export
 diffnet_to_igraph <- function(graph, slices=1:nslices(graph)) {
 
-  if (!inherits(graph, "diffnet")) stopifnot_graph(graph)
+  if (!inherits(graph, "diffnet"))
+    stopifnot_graph(graph)
 
-  pers <- graph$meta$pers
-  nper <- graph$meta$nper
+  # Filtering
+  graph <- graph[,,slices]
 
   # Listing attributes
   static.attrs <- colnames(graph$vertex.static.attrs)
   dynamic.attrs <- colnames(graph$vertex.dyn.attrs[[1]])
 
   # Creating container
-  out <- vector("list", length(slices))
-  names(out) <- pers[slices]
-
+  out <- vector("list", nslices(graph))
+  names(out) <- dimnames(graph)[[3]]
 
   for (p in 1:length(out)) {
     # Index
-    s <- slices[p]
-    dimnames(graph$graph[[s]]) <- with(graph$meta, list(ids,ids))
+    tempgraph <- graph$graph[[p]]
+    dimnames(tempgraph) <- with(graph$meta, list(ids,ids))
 
     # Graph
     tempgraph <- igraph::graph_from_adjacency_matrix(
-      graph[,,s,drop=TRUE][[1]],
+      adjmatrix = tempgraph,
       mode      = ifelse(graph$meta$undirected, "undirected", "directed"),
       weighted  = TRUE,
       diag      = graph$meta$self
       )
 
     # Computing positions
-    tempgraph <- igraph::permute(tempgraph, match(igraph::V(tempgraph)$name,
-                                                   graph$meta$ids))
+    tempgraph <- igraph::permute(
+      tempgraph, match(igraph::V(tempgraph)$name, graph$meta$ids)
+      )
 
     # Vertex Static Attributes
     for (k in static.attrs)
@@ -59,7 +60,7 @@ diffnet_to_igraph <- function(graph, slices=1:nslices(graph)) {
     # Vertex Dyn Attributes
     for (k in dynamic.attrs)
       tempgraph <-
-      igraph::set_vertex_attr(graph=tempgraph, name=k, value=graph[[k]][[s]])
+      igraph::set_vertex_attr(graph=tempgraph, name=k, value=graph[[k]][[p]])
 
     # Time of adoption
     tempgraph <- set_vertex_attr(
@@ -79,42 +80,107 @@ diffnet_to_igraph <- function(graph, slices=1:nslices(graph)) {
 #' @export
 #' @rdname igraph
 #' @param toavar Character scalar. Name of the attribute that holds the times of adoption.
-#' @param t0 Integer scalar. Passed to \code{\link{as_diffnet}}.
-#' @param t1 Integer scalar. Passed to \code{\link{as_diffnet}}.
+#' @param t0 Integer scalar. Passed to \code{\link[=diffnet-class]{new_diffnet}}.
+#' @param t1 Integer scalar. Passed to \code{\link[=diffnet-class]{new_diffnet}}.
+#' @param graph.list A list of \code{igraph} objects.
 #' @param ... Further arguments passed to \code{\link{as_diffnet}}.
 igraph_to_diffnet <- function(
-  graph, toavar,
+  graph = NULL,
+  graph.list = NULL,
+  toavar,
   t0 = NULL,
   t1 = NULL, ...) {
 
+  # At least one
+  if (!length(graph.list) & !length(graph))
+    stop("Either -graph.list- or -graph- should be provided.")
+  else if (length(graph.list) & length(graph))
+    stop("Only one of -graph.list- or -graph- should be provided.")
+
+  # Toavar cannot be missing
   if (missing(toavar))
     stop("Please provide the name of the -toa- var.")
 
-  if (igraph::is_igraph(graph)) {
-    # Getting attributes
-    toa <- igraph::vertex_attr(graph, toavar)
+  # Checking list
+  islist <- FALSE
+  if (length(graph)) {
+    if (!inherits(graph, "igraph"))
+      stop("-graph- should be of class -igraph-.")
+  } else {
+    if (!inherits(graph.list, "list"))
+      stop("-graph.list- should be a list")
+    else {
 
-    # Checking toa
-    if (!length(t0)) t0 <- min(toa, na.rm = TRUE)
-    if (!length(t1)) t1 <- max(toa, na.rm = TRUE)
+      # All must be of class igraph
+      test <- which(!sapply(graph.list, igraph::is_igraph))
+      if (length(test))
+        stop("The following elements of -graph.list- are not of class -igraph-",
+             paste(test, collapse=", "), ".")
 
-    mat <- igraph::as_adj(graph, attr="weight")
-    mat <- lapply(seq_len(t1-t0+1), function(x) mat)
-
-    vertex.static.attrs <- as.data.frame(
-      igraph::vertex.attributes(graph))
-
-    if (length(vertex.static.attrs)) {
-      test <- which(!(colnames(vertex.static.attrs) %in% c(toavar, "name")))
-      vertex.static.attrs <- vertex.static.attrs[,test,drop=FALSE]
+      # All must have the same attributes
+      test <- igraph::list.vertex.attributes(graph.list[[1]])
+      test <- which(sapply(graph.list, function(x) {
+        !all(igraph::list.vertex.attributes(x) %in% test)
+      }))
+      if (length(test))
+        stop("All -igraph- objects in -graph.list- must have the same",
+             " vertex attributes. The following differ from the first element:",
+             paste(test, collapse=", "), ".")
     }
 
-    return(as_diffnet(mat, toa, t0, t1,
-               vertex.static.attrs = vertex.static.attrs,
-               name = igraph::graph_attr(graph, "name"),
-               behavior = igraph::graph_attr(graph, "behavior"),
-               ...))
+    islist <- TRUE
   }
 
-  stopifnot_graph(graph)
+
+  # Getting attributes
+  toa <- if (!islist) igraph::vertex_attr(graph, toavar)
+  else igraph::vertex_attr(graph.list[[1]], toavar)
+
+  # Checking toa
+  if (!length(t0)) t0 <- min(toa, na.rm = TRUE)
+  if (!length(t1)) t1 <- max(toa, na.rm = TRUE)
+
+  mat <- if (!islist) igraph::as_adj(graph, attr="weight")
+  else lapply(graph.list, igraph::as_adj, attr="weight")
+
+  # Adjusting sizes
+  if (!islist)
+    mat <- lapply(seq_len(t1-t0+1), function(x) mat)
+
+  # Getting the attributes
+  vertex.static.attrs <- NULL
+  vertex.dyn.attrs    <- NULL
+  if (!islist)
+    vertex.static.attrs <- as.data.frame(
+      igraph::vertex.attributes(graph))
+  else
+    vertex.dyn.attrs <- lapply(graph.list, function(x) {
+      as.data.frame(igraph::vertex.attributes(x))
+    })
+
+  # Removing toa
+  if (length(vertex.static.attrs)) {
+    test <- which(!(colnames(vertex.static.attrs) %in% c(toavar, "name")))
+    vertex.static.attrs <- vertex.static.attrs[,test,drop=FALSE]
+  }
+
+  if (length(vertex.dyn.attrs)) {
+    vertex.dyn.attrs <- lapply(vertex.dyn.attrs, function(x) {
+      test <- which(!(colnames(x) %in% c(toavar, "name")))
+      x[,test,drop=FALSE]
+    })
+
+  }
+
+  if (!islist)
+    graph.list <- list(graph)
+
+  return(new_diffnet(mat, toa, t0, t1,
+             vertex.static.attrs = vertex.static.attrs,
+             vertex.dyn.attrs    = vertex.dyn.attrs,
+             name                = igraph::graph_attr(graph.list[[1]], "name"),
+             behavior            = igraph::graph_attr(graph.list[[1]], "behavior"),
+             self                = any(igraph::is.loop(graph.list[[1]])),
+             ...))
+
 }